requirements.txt
# Core
streamlit
pandas
numpy

# Data Visualization
plotly
matplotlib
seaborn

# Spotify Integration
spotipy
requests

# Gemini AI
google-generativeai

# Utilities
python-dotenv
pillow
cachetools
tenacity

# UI Enhancements
streamlit-option-menu
streamlit-extras

# Development
black
flake8
pytest